✨ Schema types: bootstrap, typing, codegen, docs, VS Code syntax
Made-with: Cursor

+++ compiler/codegen/testdata/golden/schema_basic.py
@@ -118,4 +118,4 @@ def _orca__invoke_tool(tool: SimpleNamespace, input_data: Any) -> Any:
 
 class vpc_data_t(BaseModel):
     region: str
-    instance_count: int
+    instance_count: float
